app/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
coloredlogs==15.0.1
filelock==3.11.0
flatbuffers==23.3.3
huggingface-hub==0.13.4
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
onnxruntime==1.14.1
packaging==23.1
protobuf==4.22.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
sympy==1.11.1
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.5.0
urllib3==1.26.15
